updated for new numpy syntax for numerical types
np.float and np.int are now depracated in favor of just float and int

+++ public_wifi/table_templates.py
@@ -8,16 +8,16 @@ import pandas as import pd
 # stars table
 columns = {
     'star_id': object,
-    'u_mast': np.float32,
-    'v_mast': np.float32
-    'star_phot_F1': np.float32,
-    'star_phot_e_F1': np.float32,
-    'star_phot_F2': np.float32,
-    'star_phot_e_F2': np.float32,
-    'star_mag_F1': np.float32,
-    'star_mag_e_F1': np.float32,
-    'star_mag_F2': np.float32,
-    'star_mag_e_F2': np.float32,
+    'u_mast': float,
+    'v_mast': float
+    'star_phot_F1': float,
+    'star_phot_e_F1': float,
+    'star_phot_F2': float,
+    'star_phot_e_F2': float,
+    'star_mag_F1': float,
+    'star_mag_e_F1': float,
+    'star_mag_F2': float,
+    'star_mag_e_F2': float,
     'clust_memb': bool,
 }
 stars_table = pd.DataFrame(columns=columns.keys())
@@ -29,15 +29,15 @@ columns = {
     'ps_exp_id': object,
     'ps_filt_id': object,
     'ps_epoch_id': object,
-    'ps_x_exp,': np.float32,
-    'ps_y_exp': np.float32,
-    'ps_u_mast': np.float32,
-    'ps_v_mast': np.float32,
-    'ps_phot': np.float32,
-    'ps_phot_e': np.float32,
-    'ps_mag': np.float32,
-    'ps_mag_e': np.float32,
-    'ps_psf_fit': np.float32,
+    'ps_x_exp,': float,
+    'ps_y_exp': float,
+    'ps_u_mast': float,
+    'ps_v_mast': float,
+    'ps_phot': float,
+    'ps_phot_e': float,
+    'ps_mag': float,
+    'ps_mag_e': float,
+    'ps_psf_fit': float,
 }
 point_sources_table = pd.DataFrame(column=columns.keys())
 
@@ -46,8 +46,8 @@ columns = {
     'stamp_id': object,
     'stamp_ps_id': object,
     'stamp_star_id': object,
-    'stamp_x_cent': np.float32,
-    'stamp_y_cent': np.float32,
+    'stamp_x_cent': float,
+    'stamp_y_cent': float,
     'stamp_path': object,
     'stamp_ref_flag': bool
 }
@@ -64,7 +64,7 @@ stamps_dq_table = pd.DataFrame(columns=columns.keys())
 # companion_status
 columns = {
     'comp_star_id': object,
-    'companion_flag': np.int16
+    'companion_flag': int
 }
 companion_status_table = pd.DataFrame(columns=columns.keys())
